Fixed incorrect handling of timeout on subscribed channel
If a subscribed channel times out the client attempts to unsubscribe.  However, in the change I made earlier the explicitly defined unsubscribe and pUnsubscribe commands did not return the status when it was called.  The effect was that when a read timeout occurs the status value is null, not equaling zero (line 698) and another read is attempted.  I personally don't think that the client should automatically unsubscribe on a read timeout favoring to just throw the exception and letting the calling function handle the read timeout.  But either way the unsubscribe commands should return the status.